% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baydis.R
\name{baydis}
\alias{baydis}
\title{Bayesian discrimination of two populations}
\usage{
baydis(X1, X2, X = NULL, R1 = 1, R2 = 1, eq.var = TRUE)
}
\arguments{
\item{X1}{a matrix or data frame that contains the explanatory variables and belongs to the first population}

\item{X2}{a matrix or data frame that contains the explanatory variables and belongs to the second population}

\item{X}{a matrix or data frame containing the explanatory variables to be classified}

\item{R1}{prior probability ratio p2/p1}

\item{R2}{the ratio of the two losses L(1|2)/L(2|1)}

\item{eq.var}{If eq.var = True, the variance is equal. Otherwise, the variances are not equal.}
}
\value{
the classification result
}
\description{
Bayesian discrimination of two populations using R
}
\examples{
\dontrun{
library(MASS)
X1 <- iris[1:50,1:4]
X2 <- iris[51:100,1:4]
res <- baydis(X1, X2)
}
}
