% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWfuction.R
\name{CFtest}
\alias{CFtest}
\title{Count Five test}
\usage{
CFtest(x, y)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{an optional (non-empty) numeric vector of data values.}
}
\value{
1 (reject H0) or 0 (do not reject H0).
}
\description{
Performs the Count Five test on vectors of data.
}
\examples{
\dontrun{
x <- rnorm(100, 0, 1)
y <- rnorm(100, 0, 0.5)
I <- CFtest(x,y)
}
}
