% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWfuction.R
\name{f.intersection_find}
\alias{f.intersection_find}
\title{The intersection points A(k) in $(0, sqrt k)$ of the curves $S_{k-1}(a)$ and $S_k(a)$}
\usage{
f.intersection_find(k)
}
\arguments{
\item{k}{k degrees of freedom}
}
\value{
a sequence with three values,the intersection point, the value of the function evaluated at it and the intersection points
}
\description{
Find the intersection points A(k) in $(0, sqrt k)$ of the curves $S_{k-1}(a)$ = P (t(k-1) > sqrt {frac{a^2(k-1)}{k-a^2}})$ and $S_k(a) = P(t(k) > frac{a^2k }{k+1-a^2})$
}
\examples{
\dontrun{
 res <- f.intersection_find(10)
}
}
