% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWfuction.R
\name{Gelman.Rubin}
\alias{Gelman.Rubin}
\title{Gelman-Rubin statistic}
\usage{
Gelman.Rubin(psi)
}
\arguments{
\item{psi}{Scalar summary statistic}
}
\value{
Gelman-Rubin statistic
}
\description{
Calculate Gelman-Rubin statistic using R
}
