% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.R
\name{lda.method}
\alias{lda.method}
\title{Linear Discriminant Analysis}
\usage{
lda.method(x, y, prior)
}
\arguments{
\item{x}{a matrix or data frame or Matrix containing the explanatory variables}

\item{y}{a factor specifying the class for each sample}

\item{prior}{prior probability for each class}
}
\value{
a list with three components:beta0 and beta1 give parameters of discrimination function and label gives the levels of classification
}
\description{
Doing linear discriminant analysis using R
}
\examples{
\dontrun{
n0 <- sample(1 : 150, 100)
iris.train <- iris[n0, ]
iris.lda <- lda.method(iris.train[, -5], iris.train[, 5],c(1, 1, 1)/3)
} 
}
