% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWfuction.R
\name{Metro.rw}
\alias{Metro.rw}
\title{A random walk Metropolis sampler for generating the standard Laplace distribution}
\usage{
Metro.rw(x0, N, sigma)
}
\arguments{
\item{x0}{initial value}

\item{N}{length of the chain}

\item{sigma}{variance}
}
\value{
a list with two components:x and acc.rate give the chain and its acceptance rate
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution using R
}
\examples{
\dontrun{
 rw<- Metro.rw(25,3000,2)
}
}
