% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.R
\name{mle.est}
\alias{mle.est}
\title{Compute mu and Sigma}
\usage{
mle.est(x, y)
}
\arguments{
\item{x}{a matrix or data frame or Matrix containing the explanatory variables}

\item{y}{a factor specifying the class for each sample}
}
\value{
a list with two components: mu and Sigma
}
\description{
Compute mle of lda/qda model, return mle of mu and Sigma using R
}
\examples{
\dontrun{
n0 <- sample(1 : 150, 100)
iris.train <- iris[n0, ]
mle.est(iris.train[, -5], iris.train[, 5])
} 
}
