% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.R
\name{one.predict.lda}
\alias{one.predict.lda}
\title{Classify one Multivariate observation by Linear Discrimination}
\usage{
one.predict.lda(beta0, beta1, x, label)
}
\arguments{
\item{beta0}{beta0 in discrimination function}

\item{beta1}{beta1 in discrimination function}

\item{x}{a vector to be classified}

\item{label}{levels of classification}
}
\value{
the classification result
}
\description{
Classify one Multivariate observation in conjunction with lda.method and also project data onto the linear discriminants
}
