% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.R
\name{one.predict.qda}
\alias{one.predict.qda}
\title{Classify one multivariate observation from Quadratic Discriminant Analysis}
\usage{
one.predict.qda(beta0, beta1, omega, x, label)
}
\arguments{
\item{beta0}{beta0 in discrimination function}

\item{beta1}{beta1 in discrimination function}

\item{omega}{the inverse of the covariance matrix*(-1/2) for each level}

\item{x}{a vector to be classified}

\item{label}{levels of classification}
}
\value{
the classification result
}
\description{
Classify one multivariate observation in conjunction with qda.method
}
