% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdfLaplaceC}
\alias{pdfLaplaceC}
\title{The density function of the standard Laplace distribution}
\usage{
pdfLaplaceC(x)
}
\arguments{
\item{x}{vector of quantiles.}
}
\value{
the density
}
\description{
Calculate density function of the standard Laplace distribution using Rcpp
}
\examples{
\dontrun{
f <- pdfLaplaceC(1)
}
}
