% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{piSugar}
\alias{piSugar}
\title{PI}
\usage{
piSugar(N)
}
\arguments{
\item{N}{the number of samples}
}
\value{
PI
}
\description{
Use Monte Carlo to calculate PI using Rcpp
}
\examples{
\dontrun{
pi <- piSugar(1000)
}
}
