% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWfuction.R
\name{psi.st}
\alias{psi.st}
\title{Scalar summary statistic}
\usage{
psi.st(X)
}
\arguments{
\item{X}{samples}
}
\value{
the scalar summary statistic
}
\description{
Calculate the scalar summary statistic using R
}
