% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.R
\name{qda.method}
\alias{qda.method}
\title{Quadratic discriminant analysis}
\usage{
qda.method(x, y, prior)
}
\arguments{
\item{x}{a matrix or data frame or Matrix containing the explanatory variables}

\item{y}{a factor specifying the class for each sample}

\item{prior}{prior probability for each class}
}
\value{
a list with four components:beta0 and beta1 give parameters of discrimination function, omega and label give the inverse of the covariance matrix*(-1/2) for each level and the levels of classification
}
\description{
Doing Quadratic discriminant analysis using R
}
\examples{
\dontrun{
n0 <- sample(1 : 150, 100)
iris.train <- iris[n0, ]
iris.qda <- qda.method(iris.train[, -5], iris.train[, 5],c(1, 1, 1)/3)
} 
}
