% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA.R
\name{qda.predict}
\alias{qda.predict}
\title{Classify from Quadratic Discriminant Analysis}
\usage{
qda.predict(struct, newx)
}
\arguments{
\item{struct}{the result of qda.method function}

\item{newx}{a matrix or data frame or Matrix to be classified}
}
\value{
the classification result
}
\description{
Classify multivariate observations in conjunction with qda.method
}
