% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWfuction.R
\name{rEk}
\alias{rEk}
\title{Random variates from the rescaled Epanechnikov kernel}
\usage{
rEk(n)
}
\arguments{
\item{n}{number of random numbers generated}
}
\value{
n random variates from the rescaled Epanechnikov kernel
}
\description{
Generate random variates from the rescaled Epanechnikov kernel using R
}
\examples{
\dontrun{
x <- rEk(1)
}
}
