% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWfuction.R
\name{rPareto}
\alias{rPareto}
\title{A random sample from Pareto(a,b)}
\usage{
rPareto(n, a, b)
}
\arguments{
\item{n}{the number of random samples}

\item{a}{a}

\item{b}{b}
}
\value{
a random sample
}
\description{
Generate random samples from Pareto(a,b) using R
}
\examples{
\dontrun{
x <- rPareto(10,1,2)
}
}
