% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetroC}
\alias{rwMetroC}
\title{A random walk Metropolis sampler for the standard Laplace distribution using Rcpp}
\usage{
rwMetroC(sigma, N, xinitial)
}
\arguments{
\item{sigma}{the variance}

\item{N}{the length of the chain}

\item{xinitial}{the initial value}
}
\value{
a list with two components:x and k give the chain and the number of candidate points rejected
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution using Rcpp
}
\examples{
\dontrun{
rw<- Metro.rw(2,3000,25)
}
}
