% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWfuction.R
\name{smsk}
\alias{smsk}
\title{The multivariate skewness statistic}
\usage{
smsk(x)
}
\arguments{
\item{x}{samples}
}
\value{
the multivariate skewness statistic
}
\description{
Calculate the multivariate skewness statistic using R
}
