\name{StatComp20060-package}
\alias{StatComp20060-package}
\alias{StatComp20060}
\docType{package}
\title{
  Final Homework
}
\description{
  It contains all the homework answers and functions for research project.
}
\details{
  Document `HWfuction`, contains all functions for homework. Document `DA` contains two algorithms, including LDA and QDA. Document `PI` contains one function and basically calculates PI using the Monte Carlo method.
}
\author{
Ma Mengyuan

Ma Mengyuan <mmy@mail.ustc.edu.cn>
}
\references{
  Hastie T , Tibshirani R , Friedman J . The elements of statistical learning. 2001[J]. Journal of the Royal Statistical Society, 2004.
}
\keyword{ package }

